+++ app/javascript/components/buda/BudaIndex.vue
